consolidate some functionality
